Move some functions to a new utils module. Rename the utils module in the tests directory to test_utils to avoid nameclashes.
